+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 
 ## v2.1 - Unreleased
 
+### Improvements
+- Add logic to alert ADMINS when Signal transport hits a CAPTCHA challenge
+
 ### Bug Fixes
 - Fix unwanted localization in badge SVG generation (#629)